Restore playable classic client and prepare shared asset workflow
Capture the recovered offline gameplay, NX integration, classic UI, skill framework, and focused regression fixtures. Include reference screenshots and complete menu-button interaction artwork.

Add the MacBook art workflow and portable button comparison. Keep generated caches, NX archives, and machine-local configuration out of this update. Validate the staged project with Windows player script compilation.

+++ Assets/Scripts/GameLogic/Tests/Core/FootholdServiceTests.cs
@@ -227,7 +227,7 @@ namespace MapleClient.GameLogic.Tests
             float slope = service.GetSlopeAt(foothold, 50);
             
             // Assert
-            float expectedSlope = System.Math.Atan2(100, 100); // Rise over run
+            float expectedSlope = (float)System.Math.Atan2(100, 100); // Rise over run
             Assert.AreEqual(expectedSlope, slope, 0.001f);
         }
